# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golf caddie backend — players, rounds, shots and course data.",
        "title": "Agentic Caddie API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/tee-holes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tee-holes"
                ],
                "summary": "Create a tee hole",
                "parameters": [
                    {
                        "description": "Tee hole to create",
                        "name": "hole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createTeeHoleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeHoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tee-holes/hole/{courseHoleId}/tee/{teeId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tee-holes"
                ],
                "summary": "Get a tee hole by course hole and tee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course Hole ID",
                        "name": "courseHoleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tee ID",
                        "name": "teeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeHoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tee hole not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tee-holes/tee/{teeId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tee-holes"
                ],
                "summary": "List tee holes for a tee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee ID",
                        "name": "teeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.teeHoleResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid tee ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tee-holes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tee-holes"
                ],
                "summary": "Get a tee hole by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.teeHoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tee hole not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tee-holes"
                ],
                "summary": "Update a tee hole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated fields",
                        "name": "hole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateTeeHoleRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tee-holes"
                ],
                "summary": "Delete a tee hole",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tee Hole ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createTeeHoleRequest": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer",
                    "example": 1
                },
                "distance": {
                    "type": "integer",
                    "example": 385
                },
                "par": {
                    "type": "integer",
                    "example": 4
                },
                "stroke_index": {
                    "type": "integer",
                    "example": 7
                },
                "tee_centre_lat": {
                    "type": "number",
                    "example": -33.8688
                },
                "tee_centre_lng": {
                    "type": "number",
                    "example": 151.2093
                },
                "tee_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.teeHoleResponse": {
            "type": "object",
            "properties": {
                "course_hole_id": {
                    "type": "integer"
                },
                "distance": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "par": {
                    "type": "integer"
                },
                "stroke_index": {
                    "type": "integer"
                },
                "tee_centre_lat": {
                    "type": "number"
                },
                "tee_centre_lng": {
                    "type": "number"
                },
                "tee_id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.updateTeeHoleRequest": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "integer",
                    "example": 385
                },
                "par": {
                    "type": "integer",
                    "example": 4
                },
                "stroke_index": {
                    "type": "integer",
                    "example": 7
                },
                "tee_centre_lat": {
                    "type": "number",
                    "example": -33.8688
                },
                "tee_centre_lng": {
                    "type": "number",
                    "example": 151.2093
                }
            }
        },
...
    }
}